requirements.txt
aiohttp
openai
python-dotenv
matplotlib
numpy
requests
smartapi-python
